+++ src/Facade/GeoFacade.php
@@ -6,11 +6,12 @@ use Illuminate\Support\Facades\Facade;
 use KMLaravel\GeographicalCalculator\Classes\Geo;
 
 /**
- * @method Geo getDistance()
  * @method Geo setPoint(array $point)
  * @method Geo setPoints(array $points)
  * @method Geo setOptions(array $options)
- * @method Geo getCenter()
+ * @method Geo getDistance($callback)
+ * @method Geo getCenter($callback)
+ * @method Geo all($callback)
  * @method Geo clearResult()
  */
 class GeoFacade extends Facade